+++ test/vue/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ["../../.eslintrc", "plugin:vue/vue3-recommended"],
+};