+++ test/vue/index.vue
@@ -1,0 +1,4 @@
+<template>
+  <div a="1" b="2" c="3" d="4">
+  </div>
+</template>