+++ ESLint.novaextension/CHANGELOG.md
@@ -9,12 +9,15 @@
 ### Changed
 
 - Performance improvements
+- Improved the "Apply a Suggestion" command
+  - Add semi-experimental choice to auto-fix all without saving to disk
+  - Slight UX changes
 
 ## v1.1.0
 
 ### Added
 
-- Apply Suggestion editor command
+- "Apply a Suggestion" editor command
 
 ## v1.0.1