+++ ESLint.novaextension/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## future
+## v1.2.0
 
 ### Added